requirements.txt
filterpy==1.4.5
lap==0.4.0
numpy==1.23.1
onnxruntime_gpu==1.12.0
opencv_python==4.8.0.74
pandas==2.0.3
Pillow==10.0.0
PyYAML==6.0.1
scipy==1.11.1
torch==2.0.1
torchreid==0.2.5
torchvision==0.15.2
ultralytics==8.0.147